\name{BAC}
\alias{BAC}
\docType{data}
\title{BAC and Fear-potenitated startle}

\description{
The \code{BAC} data frame has 96 rows and 4 columns.
The observations are the fear-potentiated startle scores
by blood alcohol concentration, trait anxiety and sex.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{BAC}{
      Blood alcohol concentration.
      }
    \item{TA}{
      Trait anxiety
      }
    \item{Sex}{
      Participant sex
      }
    \item{FPS}{
      Fear-potentiated startle
      }
    }
  }

\source{
 Loosely based on real data collected by Curtin et al from psychophysiological
 studies of alcohol effects on FPS.
}


\usage{
BAC
}

\keyword{datasets}
