\name{figLabDefaults}
\alias{figLabDefaults}
\title{Generate list graphing parameters}
\description{
  Generates a list of detailed default graphing parameters that can be used by fig functions in lmSupport for standarized graphing.  Need to use figSetDefaults with this list to save in options.
}
\usage{figLabDefaults()}
\value{
  Returns a list that includes all graphing parameters
}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}
\keyword{graphic} 