\name{varOdd}
\alias{varOdd}
\title{Tests if Number is Odd}
\description{
  Returns result of test if Number is Odd. 
}
\usage{varOdd(Number)}
\arguments{
  \item{Number}{Number to test}
}
\value{
  Returns boolean to indicate result of test
}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}
\examples{
  varOdd(3)
  varOdd(4)
}
\keyword{manip}