\name{varOdd}
\alias{varOdd}
\title{Tests if Numbers are Odd}
\description{
  Returns result of test if Numbers are Odd. 
}
\usage{varOdd(Numbers)}
\arguments{
  \item{Numbers}{Vector of numbers to test}
}
\value{
  Returns vector of booleans to indicate result of test
}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}
\examples{
  varOdd(3)
  varOdd(c(1,2,3,4,5))
}
\keyword{manip}